+++ Modules/People/Database/Migrations/2019_09_22_144730_convert_police_no_to_iteger_in_persons_table.php
@@ -1,0 +1,38 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class ConvertPoliceNoToItegerInPersonsTable extends Migration
+{
+    public function __construct()
+	{
+		// workaround for laravels limitation to change tables with an enum
+		DB::getDoctrineSchemaManager()->getDatabasePlatform()->registerDoctrineTypeMapping('enum', 'string');
+	}
+
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('persons', function (Blueprint $table) {
+            $table->unsignedInteger('police_no')->change();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('persons', function (Blueprint $table) {
+            $table->string('police_no')->change();
+        });
+    }
+}