+++ Changelog.md
@@ -9,6 +9,7 @@ Bank:
 * Only offer to create new person in bank if search results negative
 * Removed section card and registration number fields from person
 * Highlight search terms in results
+* Store police number as integer
 
 ## 2.7.6